@@ -19,7 +19,7 @@
     }
   ],
   "require": {
-    "php": ">=8.0",
+    "php": ">=8.2",
     "ext-json": "*",
     "guzzlehttp/psr7": "^2.3",
     "guzzlehttp/guzzle": "^7.4",
@@ -32,7 +32,7 @@
     "phpunit/phpunit": "^9.5",
     "vlucas/phpdotenv": "^5.4",
     "phpunit/php-code-coverage": "^9.0",
-    "friendsofphp/php-cs-fixer": "^2.16",
+    "friendsofphp/php-cs-fixer": "^3.0",
     "squizlabs/php_codesniffer": "~3.7"
   },
   "autoload": {
