@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 namespace ArangoDB\Exceptions;
 
@@ -19,7 +20,7 @@ abstract class BaseException extends \Exception
      * @param Throwable|null $previous Previous exception or error.
      * @param int $code Error code.
      */
-    public function __construct($message, Throwable $previous = null, $code = 0)
+    public function __construct(string $message, Throwable $previous = null, $code = 0)
     {
         parent::__construct($message, $code, $previous);
     }
